Added InstantMeetingScheduler.jsx and the components pertaining to it. Also changed the router to include the InstantMeetingScheduler page.
Added components:

Button.jsx: A template for a button that performs a function when clicked.

ButtonGroup.jsx: A template for 2 buttons that perform different functions, the parameters are listed as
                 onSubmit and onAddReferences.

FormField.jsx: A field template with the parameters of a regular input field.

RadioButton.jsx: A template for a radio button, used by RadioGroup

RadioGroup.jsx: A template for a list of radio buttons, can take any number of radio buttons.

InstantMeetingScheduler.jsx: The full form for the instant meeting feature, this can also be used as a template
                             for the meeting log feature, but with cancellation options.

InstantMeetingSchedule.css: The CSS for the InstantMeetingScheduler